// src/mgmt/interfaces.ts
export interface IKeypair {
    secretKey: Uint8Array;
    publicKey: Uint8Array;
    version: number | null;
}

export interface IKeypairEncrypted {
    nonce: string;
    save: string;
    version: number | null;
}

export interface ITransaction {
    inputs: ITxIn[];
    outputs: ITxOut[];
    version: number;
    druid_info: IDdeValues | null;
}

export interface ITxIn {
    previous_out: IOutPoint | null;
    script_signature: any | null;
}

export interface IOutPoint {
    t_hash: string;
    n: number;
}

export interface ITxOut {
    value: IAssetToken | IAssetReceipt;
    locktime: number;
    drs_block_hash: string | null;
    drs_tx_hash: string | null;
    script_public_key: string | null;
}

export interface IDdeValues {
    druid: string;
    participants: number;
    expectations: IDruidExpectation[];
}

export interface IDruidExpectation {
    from: string;
    to: string;
    asset: IAssetToken | IAssetReceipt;
}

export interface IDruidDroplet {
    participants: number;
    tx: { [key: string]: ITransaction };
}

export interface IAssetToken {
    Token: number;
}

export interface IAssetReceipt {
    Receipt: number;
}

export interface ICreateTxInScript {
    Pay2PkH: {
        signable_data: string;
        signature: string;
        public_key: string;
        address_version: number | null;
    };
}

export interface ICreateTxIn {
    previous_out: IOutPoint | null;
    script_signature: ICreateTxInScript | null;
}

export interface ICreateTransaction {
    inputs: ICreateTxIn[];
    outputs: ITxOut[];
    version: number;
    druid_info: IDdeValues | null;
}